Move prefs ui in its own file
Now, we have ui-prefs that handle everything related to the preferences
windows.

And we have prefs.c, which handle everything related to the
configuration file (so basically, get/set/load/save).

apply_prefs() was moved to main in the process, it has nothing to do
with the prefs window really, but it's a top-level function.